% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandwidth.R
\name{an}
\alias{an}
\title{Threshold selection for bandwidth}
\usage{
an(x)
}
\arguments{
\item{x}{matrix of observations}
}
\value{
cutoff for robust selection
}
\description{
Automated thresholds selection for the robust likelihood cross validation.
The cutoff is based on the covariance matrix of the sample data.
}
