% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{.lsum}
\alias{.lsum}
\title{Log of sum of terms}
\usage{
.lsum(lx, l.off)
}
\arguments{
\item{lx}{vector or matrix of log of terms to add}

\item{loff}{log of offset}
}
\value{
a vector of log sums
}
\description{
Computes the log of a sum of positive components, given on the log scale (\code{lx}), avoiding numerical overflow.
}
\author{
Marius Hofert, Martin Maechler (package \code{copula})
}
\keyword{internal}
