% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mig.R
\name{.mig_mom}
\alias{.mig_mom}
\title{Method of moments estimators for multivariate inverse Gaussian vectors}
\usage{
.mig_mom(x, beta, shift)
}
\arguments{
\item{x}{\code{n} by \code{d} matrix of quantiles}

\item{beta}{\code{d} vector \eqn{\boldsymbol{\beta}} defining the half-space through \eqn{\boldsymbol{\beta}^{\top}\boldsymbol{\xi}>0}}

\item{shift}{\code{d} translation for the half-space \eqn{\boldsymbol{a}}}
}
\value{
a list with components:
\itemize{
\item \code{xi}: MOM estimate of the expectation or location vector
\item \code{Omega}: MOM estimate of the scale matrix
}
}
\description{
These estimators are based on the sample mean and covariance.
}
\keyword{internal}
