% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mig.R
\name{fit_mig}
\alias{fit_mig}
\title{Fit multivariate inverse Gaussian distribution}
\usage{
fit_mig(x, beta, method = c("mle", "mom"), shift)
}
\arguments{
\item{x}{\code{n} by \code{d} matrix of quantiles}

\item{beta}{\code{d} vector \eqn{\boldsymbol{\beta}} defining the half-space through \eqn{\boldsymbol{\beta}^{\top}\boldsymbol{\xi}>0}}

\item{method}{string, one of \code{mle} for maximum likelihood estimation, or \code{mom} for method of moments.}

\item{shift}{\code{d} translation for the half-space \eqn{\boldsymbol{a}}}
}
\value{
a list with components:
\itemize{
\item \code{xi}: estimate of the expectation or location vector
\item \code{Omega}: estimate of the scale matrix
}
}
\description{
Fit multivariate inverse Gaussian distribution
}
