% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gauss_kdens_arma}
\alias{gauss_kdens_arma}
\title{Gaussian kernel density estimator}
\usage{
gauss_kdens_arma(x, newdata, Sigma, logweights, logd)
}
\arguments{
\item{x}{matrix of observations}

\item{newdata}{matrix of new observations at which to evaluated the kernel density}

\item{Sigma}{covariance matrix}

\item{logd}{logical; if \code{TRUE}, return the log density}
}
\value{
log density estimator
}
\description{
Given a data matrix over a half-space defined by \code{beta},
compute the log density of the asymmetric truncated Gaussian kernel density estimator,
taking in turn an observation as location vector.
}
\keyword{internal}
