% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gauss_lcv}
\alias{gauss_lcv}
\title{Likelihood cross-validation for Gaussian kernel density estimation}
\usage{
gauss_lcv(x, Sigma, logweights)
}
\arguments{
\item{x}{\code{n} by \code{d} matrix of observations}

\item{Sigma}{smoothing positive-definite matrix}

\item{logweights}{log vector of weights}
}
\value{
LCV criterion value
}
\description{
Likelihood cross-validation criterion function.
}
\keyword{internal}
