% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gauss_loo}
\alias{gauss_loo}
\title{Likelihood cross-validation for Gaussian kernel density estimation}
\usage{
gauss_loo(x, Sigma, logweights)
}
\arguments{
\item{x}{\code{n} by \code{d} matrix of observations}

\item{Sigma}{smoothing positive-definite matrix}

\item{logweights}{vector of log weights}
}
\value{
a vector of values for the weighted leave-one-out likelihood
}
\description{
Given a data matrix, compute the log density using leave-one-out
cross validation, taking in turn an observation as location vector
and computing the density of the resulting mixture.
}
