% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandwidth.R
\name{mig_kdens_bandwidth}
\alias{mig_kdens_bandwidth}
\title{Optimal scale matrix for MIG kernel density estimation}
\usage{
mig_kdens_bandwidth(
  x,
  beta,
  shift,
  method = c("asymptotic", "lcv"),
  type = c("isotropic", "full"),
  N = 10000L,
  pointwise = NULL
)
}
\arguments{
\item{x}{an \code{n} by \code{d} matrix of observations}

\item{beta}{\code{d} vector defining the halfspace}

\item{shift}{location vector for translating the halfspace. If missing, defaults to zero}

\item{method}{estimation criterion, either \code{asymptotic} for the expression that minimizes the asymptotic integrated squared error or \code{lcv} for likelihood (leave-one-out) cross-validation}

\item{type}{string indicating whether to compute an isotropic model or estimate the optimal scale matrix via optimization}

\item{N}{integer number of simulations to evaluate the integrals of the MISE by Monte Carlo}

\item{pointwise}{if \code{NULL}, evaluates the mean integrated squared error, otherwise a \code{d} vector to evaluate the bandwidth or scale pointwise}
}
\value{
a \code{d} by \code{d} scale matrix
}
\description{
Given an \code{n} sample from a multivariate
inverse Gaussian distribution on the halfspace defined by
\eqn{\{\boldsymbol{x} \in \mathbb{R}^d: \boldsymbol{\beta}^\top\boldsymbol{x}>0\}},
the function computes the bandwidth (\code{type="isotropic"}) or scale
matrix that minimizes the asymptotic mean integrated squared error away from the boundary.
The latter depend on the true unknown density, which is replaced using as plug-in
 a MIG distribution evaluated at the maximum likelihood estimator. The integral or the integrated
 squared error are obtained by Monte Carlo integration with \code{N} simulations
}
