% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mig_lcv}
\alias{mig_lcv}
\title{Likelihood cross-validation for MIG density estimation}
\usage{
mig_lcv(x, Omega, beta)
}
\arguments{
\item{x}{\code{n} by \code{d} matrix of observations}

\item{Omega}{\code{d} by \code{d} positive definite scale matrix \eqn{\boldsymbol{\Omega}}}

\item{beta}{\code{d} vector \eqn{\boldsymbol{\beta}} defining the half-space through \eqn{\boldsymbol{\beta}^{\top}\boldsymbol{\xi}>0}}
}
\value{
LCV criterion value
}
\description{
Likelihood cross-validation criterion function.
}
\keyword{internal}
