% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mig.R
\name{mig_mle}
\alias{mig_mle}
\title{Maximum likelihood estimation of multivariate inverse Gaussian vectors}
\usage{
mig_mle(x, beta, shift)
}
\arguments{
\item{x}{\code{n} by \code{d} matrix of quantiles}

\item{beta}{\code{d} vector \eqn{\boldsymbol{\beta}} defining the half-space through \eqn{\boldsymbol{\beta}^{\top}\boldsymbol{\xi}>0}}

\item{shift}{\code{d} translation for the half-space \eqn{\boldsymbol{a}}}
}
\value{
a list with components:
\itemize{
\item \code{xi}: MLE of the expectation or location vector
\item \code{Omega}: MLE of the scale matrix
}
}
\description{
The maximum likelihood estimators are obtained for fixed shift vector \eqn{\boldsymbol{a}}
and half-space vector \eqn{\boldsymbol{\beta}}.
}
