% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elliptical.R
\name{mle_truncgauss}
\alias{mle_truncgauss}
\title{Maximum likelihood estimation of truncated Gaussian on half space}
\usage{
mle_truncgauss(xdat, beta)
}
\arguments{
\item{xdat}{matrix of observations}

\item{beta}{vector of constraints defining the half-space}
}
\value{
a list with location vector \code{loc} and scale matrix \code{scale}
}
\description{
Given a data matrix and a vector of linear constraints for the half-space,
we compute the maximum likelihood estimator of the location and scale matrix
}
