% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandwidth.R
\name{normalrule_bandwidth}
\alias{normalrule_bandwidth}
\title{Normal bandwidth rule}
\usage{
normalrule_bandwidth(x)
}
\arguments{
\item{x}{\code{n} by \code{d} matrix of observations}
}
\value{
a \code{d} by \code{d} diagonal bandwidth matrix
}
\description{
Given an \code{n} by \code{d} matrix of observations, compute the
bandwidth according to Scott's rule.
}
\keyword{internal}
