% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandwidth.R
\name{proj_hs}
\alias{proj_hs}
\title{Orthogonal projection matrix onto the half-space}
\usage{
proj_hs(beta, inv = FALSE)
}
\arguments{
\item{beta}{vector defining the half-space}

\item{inv}{logical; if \code{TRUE}, return the inverse matrix}
}
\value{
a \code{d} by \code{d} orthogonal projection matrix
}
\description{
The orthogonal projection matrix \eqn{P} has unit determinant and
transforms an \code{n} by \code{d} matrix by taking \eqn{x * P}.
The components of the first column vector of the resulting matrix are strictly positive.
}
