% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tnorm_lcv}
\alias{tnorm_lcv}
\title{Likelihood cross-validation for truncated normal kernel density estimation}
\usage{
tnorm_lcv(x, Omega, beta)
}
\arguments{
\item{x}{\code{n} by \code{d} matrix of observations}

\item{Omega}{smoothing positive-definite matrix}

\item{beta}{vector of constraints for the half-space}
}
\value{
LCV criterion value
}
\description{
Likelihood cross-validation criterion function.
}
\keyword{internal}
