% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tnorm_loo}
\alias{tnorm_loo}
\title{Leave-one-out cross-validation for truncated Gaussian kernel density estimation}
\usage{
tnorm_loo(x, Omega, beta)
}
\arguments{
\item{x}{\code{n} by \code{d} matrix of observations}

\item{beta}{vector of constraints for the half-space}

\item{Sigma}{smoothing positive-definite matrix}
}
\value{
a vector of values for the weighted leave-one-out likelihood
}
\description{
Given a data matrix, compute the log density using leave-one-out
cross validation, taking in turn an observation as location vector
and computing the density of the resulting mixture.
}
\keyword{internal}
