% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mig_lcv}
\alias{mig_lcv}
\title{Likelihood cross-validation for kernel density estimation with MIG}
\usage{
mig_lcv(x, beta, Omega)
}
\arguments{
\item{x}{\code{n} by \code{d} matrix of quantiles}

\item{beta}{\code{d} vector \eqn{\boldsymbol{\beta}} defining the half-space through \eqn{\boldsymbol{\beta}^{\top}\boldsymbol{\xi}>0}}

\item{Omega}{\code{d} by \code{d} positive definite scale matrix \eqn{\boldsymbol{\Omega}}}
}
\value{
the value of the likelihood cross-validation criterion
}
\description{
Given a data matrix over a half-space defined by \code{beta},
compute the log density using leave-one-out cross validation,
taking in turn an observation as location vector and computing the 
density of the resulting mixture.
}
